backend_elmo_trans/requirements.txt
# FastAPI and web server
fastapi
uvicorn
pydantic

# PyTorch for CPU (Docker compatibility)
torch
torchvision
torchaudio

# Transformers and NLP
transformers
tokenizers



# Data processing
pandas
numpy
scikit-learn

# Text preprocessing
emoji
regex
stanza

# Google API dependencies
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Environment variables
python-dotenv